pyproject.toml
[tool.poetry]
name = "imdb-sentiment-analysis"
version = "0.1.0"
description = "Prompt engineering solution for the sentiment analysis task on the imdb 40k dataset."
authors = ["erik <erik.koynov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llama-cpp-python = "^0.3.8"
scikit-learn = "^1.6.1"
pandas = "^2.2.3"
tqdm = "^4.67.1"
mlflow = "^2.21.3"
dotenv = "^0.9.9"
numpy = "^2.2.4"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
jupyter = "^1.1.1"
langdetect = "^1.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
